anycast: bind to real service
The previous model was that bird bound to the anycast
healthchecker, which didn't bind to anything else.  There was a
param for the bird->anycast-healthchecker relationship, but it was
set to the same value for all users of the profile (...).

This replaces that with a fixed (in the appropriate manifest)
bound relationship from bird->anycast-healthchecker, and also
adds a new hiera param to bind anycast-healthchecker to an
underlying service daemon, and configures that to be pdns-recursor
for the dnsbox role.  This should make the stop/start sequencing
considerably more robust.

Bug: T98006
Bug: T186550
Change-Id: I80230eb06e0b794e3b28b52e468f3bc9a9917d2f