varnishmtail: add varnishttfb.mtail
Add a mtail script called varnishttfb.mtail introducing a new histogram,
varnish_frontend_origin_ttfb. The histogram tracks TTFB by origin and
cache status (hit, miss, or pass). TTFB in this context is the time it
takes varnish-fe to receive the first byte from its backend (ats-be or
varnish-be) and should be useful to compare the performance of ATS and
Varnish when acting as cache backends.

Bug: T240180
Bug: T238494
Change-Id: Iad102d8b5fd09bf2f338f6b9ebb49877b2ab1c35