cloud base images: enable passwordless login on serial0
Previously this was done on serial1; serial1 is no longer
provisioned by Nova.

This will conflate the emergency console with the console log that's
visible on Horizon.  That means that if e.g. a user types a password
on the console that password will be visible to all project members.
In most cases this doesn't violate our security policies since
project members have root on most VMs anyway, but we'll need
to make a doc change warning SREs about this potential leak
on e.g. toolforge.

Bug: T240660
Change-Id: I893f2f3576e23b8e2ba5ccaa2a167e506bc2badf