ATS: stop logging BereqURL at the TLS layer too
We have stopped logging backend request URLs on ats-be (c024ba270): this
seems like a good idea considering that the only difference between
ReqURL and BereqURL is the origin server hostname (unless there are URI
rewrites, which shouldn't happen at the ats-tls layer anyways).

This reduces significantly the the amount of data being logged per
request.

Bug: T237608
Change-Id: I0ed66789e9302733e1281c5539874a99e5f5d9a3