bacula: Setup weekly copy migrations until a first run happens
A first manual of copy job is setup on backup1001, towards
backup2001.
Configure the run to only happen every week so that the server doesn't
get overloaded with overlapping copy jobs of the first run, all trying
to copy the same jobs at the same time and resulting in a backup-dir
overload and crash, plus monitoring loss.

Bug: T238048
Change-Id: I8c15f19336ead08b106aa1aca16ac1a068eaa965