ATS: lookup cache for cookie requests
Instead of avoiding cache lookups for session/token cookie requests just
disable request coalescing.

If the request has a session/token cookie and there is a corresponding
object in cache, there are two options: either the cached object has
Vary:Cookie, or it does not.

1) If the object has Vary:Cookie, Vary-slotting makes the cache lookup a
   miss given that the request has a Cookie

2) If the object has no Vary:Cookie, it's obviously fine to return it as
   a cache hit

Due to the above, it seems safe to perform cache lookups for Cookie
requests.

Bug: T227432
Change-Id: Ibd05410fe3d8f053ee3646b52f3c7160d8330f23