cdh::hadoop: remove ipv6 constraints
What:
Remove ipv6 constraints from the hadoop package's scripts to
allow daemons to properly bind on ipv4/6 addresses.

Why:
In T225296 we discovered that /usr/lib/hadoop/libexec/hadoop-config.sh,
brought in by the hadoop package, adds a run-time parameter for
multiple Hadoop daemons: -Djava.net.preferIPv4Stack=true
This belongs to years ago when Hadoop was known to not play well with Ipv6,
something not true anymore. Recently we discovered that this file
affects also Hive, but the workaround that we are using
(adding -Djava.net.preferIPv4Stack=false at the end) is not possible
due to the way that run-time parameters are added.
This change simply removes the need for an override without
adding a patch on top of a Cloudera package.

Bug: T240255
Change-Id: Id8404a886165c843e09e6210012693a2f1ca0f46