cdh::hadoop: replace augeas with a file resource
What:
Replace the previous attempt of using augeas to remove a constraint
in a shell script (deployed via hadoop packages) with a file
resource in puppet.

Why:
augeas seems to be over complicated and difficult to manage for this
use case, a file resource is better and cleaner. The new file also
has stated at the top that it is managed via puppet, so clearer for
opsens when debugging issues.

Bug: T240255
Change-Id: I14f28c7d6194d95d68f3b90410ba1f4cd0c206bf