Move Analytics systemd timers on labstore nodes to local /mnt/hdfs
What:
Deploy on labstore100[6,7] the /mnt/hdfs FUSE mountpoint, and move
analytics systemd timers to read from it instead of remotely
(stat1007).

Why:
To ease the rollout out the hdfs-rsync tool, we thought about a interim
solution based on the /mnt/hdfs FUSE mountpoint. In this way we'll be able
to continue to pull data from Hadoop, and test hdfs-rsync properly
(like moving one rsync at the time to it rather than all in once).

Bug: T234229
Change-Id: I9e64caea78923770231923c9a27b2f32900845e9