dbctl: diffs nrpe: allow extra time to execute
We're in a situation right now where two of our transport links out of
codfw are down, which means that all codfw<->eqiad traffic is transiting
eqord, adding ~17ms of latency[0].  The conftool library is not as smart
as it could be for dbctl's use case* -- or maybe dbctl doesn't use it as
efficiently as it could -- but either way, when we execute `dbctl config
diff`, we have to make a lot (apparently 176! [1]) reads from etcd,
issued serially.  So these link outages have increased the execution
time by almost three seconds, which is enough to put us barely over the
line on NRPE's default ten seconds timeout when running in eqiad[2].

(*: Of course, there's also an argument to be made for keeping the
conftool library as simple as possible, given how important it is
to production.)

References:
[0] https://w.wiki/Dpu
[1] https://phabricator.wikimedia.org/P9869
[2] https://phabricator.wikimedia.org/P9868

Change-Id: If56e431a6115b38c2b9b51f1e23ca21be59f4fe8