toolforge-k8s: enable encryption at rest
When the hiera key is set for an encryption key, the etcd information
for secrets will be encrypted to the private key in this config.

The funny thing here is that the key is obviously available to cluster
admins. This provides a safeguard against some types of attack,
not a solution to encrypt against the world. That would require a KMS.

Bug: T240009
Change-Id: I24a958b6d45c098563bbb501e80e244bbc8d04a0