toolforge-k8s: catching init config up to live config
I just noticed this was missing in patch:
Ic655fb8530f566953fa5b69e69cdcdfbf6820349

It was quite tricky to get everything I was doing in there. This is in
the live config on all Toolforge API servers already, but a new one
won't work during a refresh without this correct.

Bug: T240009
Change-Id: I4a62d834991941ce0d7b2c19716e6bd3fdcbbd96