cloud: add ::apparmor dependency to ::profile::wmcs::mediawiki_vagrant
LXC needs apparmour to work as expected. Something has changed either in
our Cloud VPS base image build or in the packages we are applying via
Puppet that caused this dependency to be missed. Making the dependency
explicit in Puppet should correct the issues that folks are seeing when
applying ::role::labs::mediawiki_vagrant.

Change also updates the coding style of profile::wmcs::mediawiki_vagrant
to current standards.

Bug: T240875
Change-Id: I2faba681c6a458cc279e9a16fc04a8c43bd99a63