Partially revert "nova.conf ocata: remove [spice] config section"
This reverts commit 26e8ab438de8fd313b2b3b204053a410b7be5c07.

This broke so many things!  It turns out that our Stretch
image doesn't boot at all without the additional serial
port being injected by the libvirt driver, which is does
in support of the (non-existent) spice proxy.

We don't actually need any of the fancy spice-proxy things configured
here; it's enough to just enable it for nova-compute.

Bug: T240851
Bug: T240660
Change-Id: I482f5ff23182e8b59031e669374827d8426c1d7e