dynamicproxy: add backend information to access log entries
This patch modifies the access log file to include a reference about which
upstream (backend) server we are proxying the request to.

By the time of this patch, we can have 3 different backends:
 * grid engine
 * legacy k8s
 * new k8s

Having this information in the logs may helps us better understand what's going
on with requests to webservices running in Toolforge.

Example of the new log entries (shortened to make it commit msg friendly):

..labs.org 172.x.x.x - - [02/Dec/2019:..] "GET /fourohfour HTTP/[..] (backend 172.16.0.146:30000)
..labs.org 172.x.x.x - - [02/Dec/2019:..] "GET /test HTTP/1.1" [..] (backend 192.168.31.156:8000)

Bug: T238641
Change-Id: Ifd154bc992810b82f521156bd4a5d0839d47c3dc
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>