wmflib: Introduce a more usable data structure to describe services.
In its current form, lvs::configuration has become overcharged with
meaning, and it is entangled everywhere and accessed in a variety of
ways, making it nearly impossible to refactor it clean it up.

Given its complexity, it's also pretty easy to get things wrong there,
so introduce a new series of functions that allow getting the same
information in the same form from a more rationally-constructed set of
functions. Once the transition is completed, only these functions should
be used to access said data.

Also, this patch introduces some additional info in the services
details, including the state in a sort-of state machine for the
introduction of a service, and if the service is encrypted or not.

Change-Id: I3f979143f3965f49fe155381810196e37ddfa8d3