varnish: Remove duplicate 'Content-Type: text/html' statement
It works fine and I haven't seen any issue with it, I just noticed
in review that in the changed file and in browsersec.vcl.erb, the same
statement is run. The subroutine isn't used anywhere else, either.

Could go either way, but I've left the one closer to the output
in place and removed the dupe in the caller. This way, if it gets
used elsewhere, it will also get the correct Content-Type no matter
what.

Change-Id: Ic668e4f6ee1a14b38acb71ccf92c9a95130c9589