mediawiki: Capture shutdown/destruct backtrace in php7-fatal-error.php
Also documented two examples. I'm not sure if there is a deeper
source that we might be able to influence to make this better in
the first place, but seems easy enough to handle for now.

The previous regular expression this code was using does not appear
to be matching against anything I've seen from PHP 7.2 in our current
setup.

Bug: T241097
Change-Id: Id9fe711b5d60deace5babd767b4592611bbd472b