Switch to digicert-2019a for eqsin and esams
Revocation issues are over with and these certs are safe to use
again, so (re-)deploy them at the non-US edges.  This will cause
some minor perf boosts due to the smaller size of Digicert's
stapled OCSP responses and the shorter chain to the level of root
compatibility we require.

Bug: T238494
Change-Id: Id02606b5d3ea2c9465e8127c4ce36d6cdd3b4654