puppet-merge: ensure puppet is only run from one server
If puppet-merge is run from two servers simultaneously there is a race
conditions which means each puppet run will only succeed on some
systems.  This updated forces everyone to use the same server so we can
ensure the locking works.  This rule will be removed once we implement
distributed locking

Change-Id: I7199942abeecedc5b5c40303c754297c7ae23ca9