labs: send integration alerts to the team
I receive Icinga alarms for the 'integration' WMCS project, but I do not
scale. In the interest of sharing the burden among the team, send the
notifications to the release engineering mailling list.

There are only a few notifications per week, usually about puppet/apt
issues and sometime about a disk partition that has filled up.

Change-Id: Ic17dee1ce23d60dd912148f87d4c3be80535db79