puppet-merge: it isn't lack of diffs that matters
In the event that an empty patch was submitted, the tree hash
is still different, of course.  This does happen from time to time
because of simultaneous identical bugfixes (e.g. 2c64210).

So let's just care only about tree hash, instead of using presence or
absence of diffs to make any decisions.

Change-Id: I2ce3abf7793def10a9f10e0fca858deafd3dcb78