nova firstboot script: disable 'growpart' in cloud-config
Previously we passed in a simple bash script to cloud-init; now
we pass in a mime multi-part file instead.  The multipart file
contains the old firstboot script and also a cloud-config
yaml block.

The cloud-config block disables the default partition-growing
behavior, leaving us with our standard, small root volume
so that the rest of the space can be claimed by lvm.

Bug: T241322
Change-Id: Ie6ff383941090944f4ab3b187c08e7d4fa9d4c66