Add a define to install a package from a repository component
This define can be used to install a package from a component provided on
an internal apt repository. Correct ordering is ensured so that repositories
are added, apt refreshed before the packages are installed and pinning
is setup for apt.

This should avoid DRY and prevent a number of error cases which can happen
if adding a repository component incorrectly.

Bug: T240324
Change-Id: I013e0e861433d1428730c369ed545320745138da