openstack: reduce again number of DB connections
Apparently openstack is designed to take over a database completely, and in our
current database for the eqiad1 deployment, m5-master, we have limits that
result in some openstack components malfunctioning. This in turn results in
page alerts in the middle of the holidays :-(

That was the case today. Try to further reduce number of DB connections again,
for nova, keystone and neutron.

Change-Id: I3a4e8c1a3951f853c77aa272fc88015e6fe66e3e
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>