wdqs: use RecentChanges API for updates on all WDQS servers
Since kafka stream relies on chronology_id and we might need to remove
it, let's switch to RecentChanges for updates, at least temporarily.

This activate RecentChanges on all wdqs servers and should be merged
only once the change has been validated on wdqs1009.

Note: there are various minor performance improvements that have been
implemented only on the kafka updater, there is a minor risk of
performance degradation with this change.

Bug: T241410
Change-Id: I1c0b8830618f1d7bbd56dd3a18cbf09f6fdad282