neutron: update l3_agent_hacks for Pike
Introduce modifications to config.py and router_info.py to support our custom
options dmz_cidr and routing_source_ip, which we use in CloudVPS for our
networking architecture.

This patch updates the 'original/' files to match those from the Pike
upstream, as well as the patched files.  The patched files were
generated by diffing in the 'ocata' subdir, then applying
that patch to the 'pike' upstream files.

The ocata patch can be seen with the following commands:

 diff -u modules/openstack/files/ocata/neutron/l3/original/router_info.original \
         modules/openstack/files/ocata/neutron/l3/router_info.py

 diff -u modules/openstack/files/ocata/neutron/l3/original/config.original \
         modules/openstack/files/ocata/neutron/l3/config.py

The diff contained in this very patch can be generated anytime with the
following commands:

 diff -u modules/openstack/files/pike/neutron/l3/original/router_info.py.original \
         modules/openstack/files/pike/neutron/l3/router_info.py

 diff -u modules/openstack/files/pike/neutron/l3/original/config.py.original \
         modules/openstack/files/pike/neutron/l3/config.py

Bug: T241347
Change-Id: I07e426859ff97414497c386fd630ca6f9e59a4c5