neutron pike: roll back our router_info.py slightly
The 'original' file I imported previously was from the 'pike' git
branch.  It turns out that the git pike branch is slightly ahead
of the branch that is actually packaged for debian.

This patch updates original/router_info.py to the version that
actually ships for debian, and updates our hacked file accordingly.

config.py is the same in git and debian, so no need for
change there.

Bug: T241347
Change-Id: Id5f0871a33896cbc3cf1fa2805efad81f046d158