wdqs: enable async_import on eqiad public cluster
The recent switch to RecentChange for polling updates on WDQS has
some performance impact. wdqs1004 on which async_import was being
tested does not see any significant lag, so it appears that this
would mitigate the issue.

This patch enables async_import for the eqiad public wdqs cluster,
which is the only one that sees significant user generated load
and thus the only one having significant update lag.

Bug: T241410
Bug: T241536
Change-Id: I898c5ddfaf15a34495316bf45dbda47e431e4877