wiki replicas: Remove outdated comment about spamblacklist
In October 2017 (T178052) spamblacklist log entries were considered
privileged information. That opinion changed in January 2018 (T184483)
when spamblacklist log entries were made visible through Special:Log and
the Wiki Replica views. We neglected to remove the comment from 2017
when the log type was added to the allow list in 2018 which is causing
confusion now.

Bug: T241668
Change-Id: I28377bfdeb1f9dc0087b5ba23584169a4cd6b444