toolforge: Monitor local crontabs with Prometheus
Replace our legacy diamond collector for crontab files with a Prometheus
node_exporter compatible collector. The Prometheus implementation does
not include the level of configuration possible with the prior Diamond
collector. Instead it hard codes the needed functionality and can be
extended to support additional used cases in the future when they are
found.

This changeset removes the old collector and adds the new collector in
a single change. Grafana dashboards and other systems monitoring the
legacy graphite metrics will need to be updated soon after deployment.

A lint ignore block has been added to the ::toollabs class for the cross
module import of profile::prometheus::node_local_crontabs. This legacy
class and module will be removed "soon". It is currently blocked on the
decommission of the legacy Kubernetes cluster in Toolforge.

Bug: T210993
Change-Id: I0a597b2daab5107155d16ea138e36e863a09e2f9