Switch DNS servers and contemporary LVSes to standard Partman recipes
This will create the following on next reimage:

 * GPT partitions:
   - boot 300MB (biosgrub type, see below)
   - LVM
 * LVs:
   - /:    ext4, 80GB
   - /srv: ext4, up to 80% of the VG (on Buster, see below)
   - swap: swap, 1GB
   - rest: unallocated (on Buster, see below)

(The partman-auto-lvm/guided_size setting is supported starting with Buster.
If this configuration is used on Stretch and earlier there will be no
space left free on the VG, and the last LV (/srv) will take up all the space
in the VG)

The LVSes don't need an /srv, but it still seems useful to apply the
new standard layout nonetheless.

Bug: T156955
Change-Id: I9f4cf58929ddb65332394ac25faf254c15689c80