toolforge: prometheus: fix label config for cadvisor metrics
We want to fetch each cadvisor pod metrics individually. Each cadvisor pod is
deployed in a DaemonSet in every worker node, so this means we would be
fetching metrics from every worker node, which is exactly what we want.

Before this patch, prometheus was collecting metrics from a service which would
be RR load-balanced between the different pods, so prometheus would see
inconsistent metrics between scrapes.

In this patch, while at it, drop the custom service we had for cadvisor, we no
longer need it.

Bug: T237643
Change-Id: I4d3c877a29e2696e69b40b287c9d07d8b1ae8834
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>