Deprecate raid1-lvm-ext4-srv-dualboot.cfg
Switch thumbor* and theemin to the new standard Partman recipes. The new
recipes write grub-installer/bootdev to all devices and set up the following
disk layout:

* GPT partitions:
  - boot 300MB (biosgrub type, see below)
  - LVM
* LVs:
  - /:    ext4, 80GB
  - /srv: ext4, up to 80% of the VG (on Buster, see below)
  - swap: swap, 1GB
  - rest: unallocated (on Buster, see below)

The partman-auto-lvm/guided_size setting is supported starting with Buster.
If this configuration is used on Stretch and earlier there will be no
space left free on the VG, and the last LV (/srv) will take up all the space
in the VG.

Bug: T156955
Change-Id: I91f5242f93357698edf290b10ca979ba729055ea