graphite::wmcs::archiver: move warnings to /dev/null
What:
Move stderr output of the archive-instances script to /dev/null.

Why:
This is the last step of a series of patches aimed to avoid daily
cronspam from cloudmetrics100[1,2] hosts. The script seems now
working fine on both nodes, but it emits this on a daily basis:

/usr/lib/python2.7/dist-packages/keystoneauth1/adapter.py:200:
  UserWarning: Using keystoneclient sessions has been deprecated.
  Please update your software to use keystoneauth1.
  warnings.warn('Using keystoneclient sessions has been deprecated. '

The proper solution should be to either move the cron to a systemd
timer (avoiding the emails to root@) or to upgrade the archive-instances
script to avoid the warning. In the meantime, it is good to remove
cronspam to root@.

Change-Id: I561b0e4309d1da22ac51f7e092f6e52c0760a3db