airflow: Remove PrivateTmp from airflow-kerberos unit
The airflow-kerberos unit has the singular job of maintaining
up-to-date credentials in a /tmp directory. This needs to be read in
by the airflow-scheduler unit, along with admins running
`airflow test ...` commands. Ensure this file is visible where necessary.

airflow-webserver does not interact directly with kerberos, and as
such keeps its PrivateTmp configuration. airflow-scheduler was already
configured without PrivateTmp.

Change-Id: I031f3f313082cf2dfea55cd4cf09fd7aceb32615