airflow: Share kerberos cache via /var/run
The airflow-kerberos unit has the singular job of maintaining
up-to-date credentials in a /tmp directory. This needs to be read in
by the airflow-scheduler unit, along with admins running
`airflow test ...` commands. Ensure this file is visible where necessary.

PrivateTmp was preventing these from being shared. Move the shared
file into /var/run to make the sharing explicitly allowed.

Change-Id: I70ced75d05762eda6a9b1d8d3f036f02ea43cd1d