hive: store delegation tokens in the db
What:
Store Hive's delegation tokens in a database, as opposed to in
memory, to be resilient on reboots/restarts/etc.. and avoid
issues with Hue and Oozie.

Why:
Several users reported issues while using the Hive tab in Hue,
that seem to be related to Hue not being able to renew delegation
tokens as Hive expects. As workaround, the following document suggests
to persist hive delegation tokens on the db, rather than in memory (default):
https://mapr.com/support/s/article/Intermittent-Hue-Hiv-connectivity-issue-in-MapR-SASL-secure-cluster?language=en_US
Other reports seem to indicate that Oozie may end up in the same situation
as well, so this fix might help T241650 as well.
The change has been tested in Hadoop test, the issue was reproduced but not
after adding the new setting.

Bug: T238560
Change-Id: I39ca5fee22116c524ad1df086d316a4bd0cdf41c