Deprecate raid1.cfg
The eventlog* part of the config is actually moot, only eventlog1002 is left
and it's already matched above with a dedicated entry.

The scb servers will also very likely not be reimaged again, but if that
should ever happen for some reason the standard recipe is an adequate
replacement (which uses ext4 instead of XFS as the major difference).

Bug: T156955
Change-Id: I58d7a112776bd51040382231a822caffcf41b8e5