motd: fix awk command for parsing puppet configuration_version
If the configuration variable (commit message) contains speech marks then
puppet will not dump the string surrounded by speech marks as such we
need to parse out the string in a different manner.

This update uses ruby yaml to parse out the value.  We first grep the
bit of yaml we are interested in to save is loading the entire summery
file which can take some time.

Bug: T241459
Change-Id: Id2441e07729cdeba522ead54effe0dcb2a50403f