apt::package_from_component: Switch to a per component apt-get exec
Re-using the global "apt-get update" exec from the apt module didn't
work out in a test reimage of a video scaler; per Puppet logs a notification
to the exec was triggered, but the subsequent package installation used
the default apt config. Switch to a component-specific exec instead, it
adds a little overhead during initial reimage, but should work reliably.

Change-Id: I14e9bc7ce338f9318babf5469042d862c0dc7ee2