Enable hive kerberos connections from search/airflow instances
When spark-submit initializes an application it connects to the
metastore and provides a delegation token to the app it starts up.
Currently an-airflow1001 is unable to connect to metastore with either
spark or the `beeline` cli app. Adds kerberos config for hive matching
config for other profiles in the analytics network

Change-Id: I0dbf47a16254a0284799749cae3d274aa658a073