gerrit: adjust bacula backup behaviour to deal with multiple hosts
We don't want to tie "is_replica" directly to the question if we
want Bacula backups for a Gerrit host or not.

If we have more than 2 servers, like the temp. test server, we want
to be able to switch it on/off separately.

Backups of test servers would be a waste.

Bug: T239151
Change-Id: I66c09ac24037ff5d0ea9b0665a10e7d606b0537b