airflow: Provide wrapper script to invoke airflow
In multiple circumstances such as the airflow webserver invoking
gunicorn or the scheduler starting workers, airflow is shelling out
and expecting to find the bin/ dir of the virtualenv on the path.

Ensure that the airflow command is always executed in the context of
the virtualenv. Additionally take this opportunity to centralize
setting AIRFLOW_HOME to keep things dry.

Change-Id: I5c4fd6f16c89a06c58d42947565be16c1464d583