toolforge: new k8s: prometheus: scrape metrics from each individual ingress pod
The current scrape configuration is to use an intermediate service. When
prometheus hits this service, the request is directed to an arbitrary
ingress-nginx pod. This results in prometheus having inconsistent metrics about
what ingress-nginx is doing.

Instead of this, instruct prometheus to discover each nginx-ingress pod and
scrape each one using the k8s API url proxy for the pod.

While at it, drop the nginx-ingress-metrics k8s service, no longer required.

Bug: T237643
Change-Id: Ibe17b2111f5dfe181a1713ef6cd6505d59ade01c
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>