Add role to mc-gp200[1-3]
What:
Add a new role mediawiki::memcached::gutter to mc-gp200[1-3].

Why:
The new hosts mc-gp200[1-3] (and sooon mc-gp100[1-3]) are the memcached
hosts in codfw dedicated to the mcrouter's gutter/failover pool.
They run Debian Buster and memcached 1.5 (as opposed to 1.4 in Jessie),
and they are beefier (256G of RAM) so they need different parameters.

Bug: T239249
Change-Id: I53ec1946711b1eb1bd88240fe58d49518e04d22a