labs prometheus: only bind localhost and update vhost config
Configure the labs prometheus server to only bind to the loopback interface. All
requests to the server should come through the local Apache proxy.

Set the prometheus vhost server as the default Apache vhost. The cloudmetrics
hosts have multiple vhosts and our internal tooling expects prometheus to be the
default.

Bug: T242460
Change-Id: I293e1260161984e21f11cfe2f0e2a5c4e0b70d64