lvs::monitor: use unique identifiers for services
We have for a long time had the inability to declare multiple
monitoring services for the same lvs hostname.  I7fcc7db7661ff663
partially solved the issue, while maintaining mostly resource-name
compatibility with the previous status-quo. This patch completes that
work so that now we can declare an arbitrary number of services per hostname.

Change-Id: Ia9a0049295ecb29448098c5d9d98aca89494b181