Deploy Apache BigTop's apt repository on analytics1031
What:
Swap Cloudera's CDH apt settings with BigTop's ones on analytics1031
as test (the host is part of the Hadoop cluster).

Why:
This test allows us to establish if a Hadoop worker with BigTop
packages can work in a CDH cluster. It will also tell us if any
package is missing from BigTop's repository.

Change-Id: Ic9332df17ba12e0f4601c5fd834d2bfe7746a0cb