Enable Kerberos for Analytics Refine Spark jobs on the Hadoop coords
What:
Enable the Kerberos support for Refine Spark jobs that interact with HDFS.

Why:
This option was supposed to be already enabled, but when turning on Kerberos
we forgot to do it. When we enable kerberos in spark jobs, what really happens
is that the spark-submit command is executed via kerberos-run-command, that
populates the Kerberos analytics user's credential cache to allow Spark to pick
credentials up when starting. Since on the Hadoop coordinators other jobs are
already populating the analytics' credentials cache, it has been working so far.
We received a failure related to a Hive CREATE table statement (for a new event)
that we know requires spark to pass a keytab to the Yarn's cache to be used later
on.

Change-Id: I0c8148b9961396887ecc88fb79ed65e619bb8912