Clean labstore fetcher after hdfs-rsync move
Labstore fetcher was using HDFS fuse mount to copy data from hdfs
to the local folder serving those files externally.
In previous patches we have converted the jobs to use hdfs-rsync,
a JVM tool using HDFS API to copy files from/to hdfs from/to local
with rsync advatages of not copying not changed files.
This patch cleans the leftovers of the move, noticeably removing
old jobs (after having marked them as absent for a while), and
removing unneeded code and variables from the relevant classes.

Change-Id: I03fffa55561e3f71d2965a5cbdf685c1dfe1d6ae