dynamicproxy: urlproxy: introduce support for domain-based routing
Modify urlproxy LUA code to evaluate first the subdomain of the request. We are
interested in the left-most subdomain, it might indicate which tool the user
is interested in. If we find a match in redis, then proceed selecting a
backend as normal.
If we don't find a backend using the subdomain, then evaluate the rest of the
URL as we were doing before (another redis lookup).

Doing it this way, we give support for the 2 addressing schemes:
 * $toolname.toolforge.org
 * tools.wmflabs.org/$toolname

This patch is specifically about giving support for the legacy gridengine
webservice setup + the legacy kubernetes cluster. Mind that webservices running
in the new kubernetes cluster don't register in redis at all.
Once all webservices are happy with the new addressing scheme, we could
consider dropping the code for the old scheme.

I'm aware this is kind of mixing urlproxy and domainproxy functionalities.

While at it, prevent nginx from hijacking request to simple / as redirections
to the admin tool. That is something that the fourohfour tool will need to
handle.

Bug: T234617
Change-Id: I1f48e239a394dbb36e2c93f69bc19b2e06ad4b5e
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>