conftool::safe_service_restarts: better support for non lvs servers
Sometimes it comes in handy to install the same script on one
load-balanced cluster and on a single separate server. This is even
more true across the production / deployment-prep boundary.

This patch tries to fix some of the technical debt around that and
also the mediawiki::php::restarts profile.

With this patch:
* If has_lvs is false, the restart script will work correctly without
requiring profile::lvs::realserver or actually using conftool
* As a consequence, no stub services declaration will be needed in
deployment-prep.

Change-Id: I6439df2b3b567fbff589081e5903a5ffa25d908d