airflow: Provide runtime directory for skein
Skein starts up a daemon that acts as the intermediary between python
and the yarn cluster and stores credentials in SKEIN_CONFIG. This
defaults to ~/.skein, but the analytics-search user does not have a
home directory so this fails. Set it to an explicit path that we
know is writable.

Change-Id: I321f98473855bcb5157a9174bc4c4c12136ccc5e