Switch some of the WMCS systems to standardized Partman recipes
We are standardizing our Partman recipes. This will use the following
layout on next reimage:

* GPT partitions:
  - boot 300MB (biosgrub type, see below)
  - LVM
* LVs:
  - /:    ext4, 80GB
  - /srv: ext4, up to 80% of the VG (on Buster, see below)
  - swap: swap, 1GB
  - rest: unallocated (on Buster, see below)

The partman-auto-lvm/guided_size setting is supported starting with Buster.
If this configuration is used on Stretch and earlier there will be no
space left free on the VG, and the last LV (/srv) will take up all the space
in the VG.

Bug: T156955
Change-Id: I0003d46d3769fe24010e2be6f1ed0d1aa07231e1