profile::analytics::cluster::packages::common: add libcrypto.so link
What:
Add the libcrypto.so symlink to /usr/lib/x86_64-linux-gnu.

Why:
Hadoop needs a non versione libcrypto.so link in the path to
be able to load openssl's functionalities. In Stretch the link
is brought by libssl-dev, but it links to libcrypto1.1.0 that
is not compatible with Hadoop.

The overall goal is to allow Spark RPC encryption (leveraging
openssl's AES algs) that currently fails intermittently with
Java JCE.

Bug: T240934
Change-Id: I7a5d3f199334c6da0891c3551961bdb0be70d472