service::catalog: fix host header for check_http* monitoring
Right now in many stanzas we're declaring as monitoring check_command
the url of the service with "svc.%{::site}.wmnet" as Host
header. Given hiera gets interpreted at compile time, %{::site} would
be eqiad or codfw, depending on which icinga server we're compiling
for.

This would result in working but "incorrect" checks like
serviceA:
(eqiad) -> Host serviceA.svc.eqiad.wmnet
(codfw) -> Host serviceA.svc.eqiad.wmnet

on the eqiad icinga host, and the opposite on the codfw icinga host.
While this usually works, it is incorrect. So, switch to using the
discovery url whenever appropriate, and in general stop using
%{::site} in monitoring sections.

Change-Id: I24d1d57ed0cda4590cb6914a7a7e83f087a64898