profile::kerberos: check the principal before creating it
What:
Add a check in manage_principals.py to avoid to execute the
commands to create a principal if already created.

Why:
It happened to me that I accidentally executed the last command
saved in my bash history to create a principal, and the command
executed successfully. It is very confusing for users to get another
email related about a brand new account.

Change-Id: I8111ccc44515d55df8313f8c6ad350c4107a4ffe