toolforge: toolviews: don't run it if not in the tools CloudVPS project
We are not interested in running toolviews if not in the tools CloudVPS
project.

This happens specifically in toolsbeta, where we have proxy servers, they have
this manifest deployed but credentials to connect to the DB are wrong, thus
producing cronspam. Right now we are not interested in data from toolsbeta
anyway, only actual tools running in Toolforge.

Change-Id: Iee4b223c0d140f975059099e5520b6607da10405
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>