realm: introduce global variable $wmcs_deployment
This variable holds the openstack deployment name. It is meant to be used by
virtual servers (VM instances) running inside our openstack deployments.

To generate the value, we do some computing on the machine FQDN as we have been
doing for the $labsproject variable in the past.

This patch also adds support for our new domain names that we plan to introduce
"soon" (already introduced in codfw1dev):

Current/legacy FQDN for VMs:
 $vmname.$project.eqiad.wmflabs

New FQDN for VMs:
 $vmname.$project.$deployment.wikimedia.cloud

Bug: T242607
Change-Id: I914d163123a350e625daacb5de38479d0e0d99da
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>