Fix log spam from arclamp-generate-svgs
The list of log files to process is generated at the top of a
potentially long-running loop.  If the log file is no longer available
between when the shell glob runs and when the loop gets around to
processing the file in question, an error is emitted.  We should skip
these inputs instead.

Bug: T243762
Change-Id: Ie548828760f047c936fc21506326296827cf829b