CloudVPS: set domain for VM instances using nova-api metadata
This way we can easily configure the domain based on hiera configuration (i.e,
per deployment domain). We already have the hiera keys in place, (dhcp_domain)
so this patch just reuses them.

We aim to support 2 different domains for VM instances:
 * eqiad.wmflabs               -- current domain, eventually the legacy one
 * $deployment.wikimedia.cloud -- the new one

From the VM point of view, this shows up like this:

=== 8< ===
 root@bastion-codfw1dev-01:~# curl http://169.254.169.254/openstack/latest/vendor_data.json
 {"domain": "codfw1dev.wikimedia.cloud"}

 root@bastion-codfw1dev-01:~# curl http://169.254.169.254/openstack/latest/vendor_data.json \
                              | sed -r 's/^.*domain\": \"//' | sed -r 's/\"}//g'
 codfw1dev.wikimedia.cloud
=== 8< ===

This should help improve a bit the firstboot mechanism. Only do a basic
healthcheck in the script.

Note to self: this script is only loaded at instance creation time, i.e., to
live-test it you need to create a new VM instance.

Bug: T243556
Change-Id: I8e905c282d81be588f605204eb1e0f6c016e464c
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>