hieradata: openstack: introduce basic keys for codfw1dev
We need these basic keys for codfw1dev. The eqiad1 counterparts are in
hieradata/labs.yaml but I don't like that file and would like to avoid adding
more data to it until we fix the cloud at puppet.git hiera mess for real.

While at it, fix the TLD and domain that ends up in /etc/resolv.conf.

Bug: T243556
Change-Id: I61875dc77485bb527fd64ccca0fb0e6e7eb8432c
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>