cloud: hiera: puppetmaster: refactor hiera (for VM instances)
This patch introduces a major refactor/rearrange into how hiera works inside
CloudVPS. The problem this is trying to solve is our current difficulties to
support multiple openstack deployments, in the sense of not having a clean,
consistent and robust mechanism for storing hiera data per deployment.

By the time of this patch we have 2 openstack deployments:
 * eqiad1 -- primary deployment, this is the main CloudVPS service
 * codfw1dev -- development/testing/stagging deployment

(but please note that not so long ago we had 4 different deployments).

The differences in the setup between eqiad1 and codfw1dev makes it a bit
difficult to actually use codfw1dev as testing/stagging environment. There are
things we simply cannot test there because limitations in the current setup due
to how hiera is done.

Previous to this patch, there was a common 'hieradata/labs.yaml' file with all
sorts of hiera configuration that gets applied to both deployments arbitrarily.
This is confusing, and something we can improve. The hierarchy tree (as seen
by the puppetmaster) is also not consistend between the 2 deployments.

Andrew had the idea of introducing %{::site}/codfw1dev-instances in the search
tree, and this patch is digging depper on that idea of having per-deployment
hiera configuration.
Another win of this patch is that we get rid of yet another 'labs' keyword,
which may eventually lead to rework the entire 'labs' realm as a concept.

In general, this new organization scheme should help us have arbitrary
deployments.

Meaning of new data files:
* hieradata/cloud.yaml
  to hold general hiera keys for any VM in any deployment
* hieradata/cloud/<deployment>.yaml
  to hold general hiera keys for any VM in a given deployment
* hieradata/cloud/<deployment>/<project>/common.yaml
  to hold general hiera keys for any VM in a given deployment/project
  This is the equivalent of using the Horizon puppet project setting
* hieradata/cloud/<deployment>/<project>/hosts/<hostname>.yaml
  to holdhiera keys for a particular VM in a given deployment/project
  This is the equivalent of using Horizon puppet VM setting

Still, in general, horizon settings are searched first.

A similar patch will follow in the private repos.

Bug: T229441
Change-Id: Id7c659c26652021a9caf85285f3a474e2269c163
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>